fix(tests): make tests work with tag expr_noreflectmethod
If all tests of a file are failing, just add a build constraint. If only
some tests are failing, skip them with the SkipNoReflectMethod helper.
Failing examples are moved into a dedicated test file with a build
constraint.
