+++ test/issues/756/issue_test.go
@@ -1,3 +1,5 @@
+//go:build !expr_noreflectmethod
+
 package issue_test
 
 import (
