+++ test/issues/844/issue_test.go
@@ -1,3 +1,5 @@
+//go:build !expr_noreflectmethod
+
 package main
 
 import (
