+++ test/operator/issues584/issues584_test.go
@@ -1,3 +1,5 @@
+//go:build !expr_noreflectmethod
+
 package issues584_test
 
 import (
